# swagger_ui/openapi.yml
openapi: 3.0.2
info:
  title: Redis-REST with Referencing for Node normalization
  description: 'A REST interface for a Redis database with referencing for Node normalization.</br></br>
  The data in this compendium is gathered from the following sources:
  </br> - PubMed (PMID)
  </br> - Etc.'

  version: '1.0.0'
  contact:
    email: kebedey@renci.org
  termsOfService: http://robokop.renci.org:7055/tos?service_long=Redis-REST+with+Referencing&provider_long=the+Translator+Consortium
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
tags:
  - name: Interfaces
paths:
  /get_semantic_types:
    get:
      tags:
        - Interfaces
      summary: 'Get the comprehensive list of semantic types.'
      description: 'Returns a distinct set of the semantic types discovered in the compendium data.'

      responses:
        '200':
          description: Results
          content:
            application/json:
              schema:
                type: object
                example:
                  semantic_types:
                    types:
                      - "cellular_component"
                      - "named_thing"
                      - "etc."

  /get_curie_prefixes:
    get:
      tags:
        - Interfaces
      summary: 'Get the curie prefix(es) for the semantic type(s) entered.'
      description: 'Returns the curies for a semantic type(s).'
      parameters:
        - in: query
          name: semantictype
          schema:
            type: array
            items:
              type: string
            example:
              - "chemical_substance"
              - "biological_entity"

      responses:
        '200':
          description: Results
          content:
            application/json:
              schema:
                type: object
                example:
                  chemical_substance:
                    curie_prefix:
                      - "GO"
                      - "PUBCHEM"
                      - "etc."

                  biological_entity:
                    curie_prefix:
                      - "CHEBI"
                      - "UNICHEM"
                      - "etc."

  /get_normalized_nodes:
    get:
      tags:
        - Interfaces
      summary: 'Get the equivalent identifiers and semantic types for the curie(s) entered.'
      description: 'Returns the equivalent identifiers and semantic types for the curie(s) entered.'
      parameters:
        - in: query
          name: curie
          schema:
            type: array
            items:
              type: string
            example:
              - "GO:0032157"
              - "UMLS:C0856524"

      responses:
        '200':
          description: Results
          content:
            application/json:
              schema:
                type: object
                example:
                  GO:0032157:
                    id:
                      identifier: "GO:0032157"
                      label: "prospore contractile ring"
                    equivalent_identifiers:
                      - identifier: "GO:0032157"
                        label: "prospore contractile ring"
                    type:
                      - "cellular_component"
                      - "named_thing"
                      - "etc."

                  UMLS:C0856524:
                   id:
                    identifier: "UMLS:C0856524"
                  equivalent_identifiers:
                   - identifier: "UMLS:C0856524"
                   - identifier: "MEDDRA:10035322"
                  type:
                  - "phenotypic_feature"
                  - "named_thing"
                  - "etc."